automatically apply OneVsAll in DecisionTree(renamed from DecisionTreeTrainingModel)
