+++ NeoML/docs/en/API/ClassificationAndRegression/DecisionTree.md
@@ -1,8 +1,8 @@
-# Decision Tree Classifier CDecisionTreeTrainingModel
+# Decision Tree Classifier CDecisionTree
 
 <!-- TOC -->
 
-- [Decision Tree Classifier CDecisionTreeTrainingModel](#decision-tree-classifier-cdecisiontreetrainingmodel)
+- [Decision Tree Classifier CDecisionTree](#decision-tree-classifier-cdecisiontree)
 	- [Training settings](#training-settings)
 	- [Model](#model)
 	- [Sample](#sample)
@@ -11,11 +11,11 @@
 
 Decision tree is a classification method that involves comparing the object features with a set of threshold values; the result tells us to move to one of the children nodes. Once we reach a leaf node we assign the object to the class this node represents.
 
-In **NeoML** library this method is implemented by the  `CDecisionTreeTrainingModel` class. It exposes a `Train` method that allows you to train a multi-class classification model.
+In **NeoML** library this method is implemented by the  `CDecisionTree` class. It exposes a `Train` method that allows you to train a multi-class classification model.
 
 ## Training settings
 
-The parameters are represented by a `CDecisionTreeTrainingModel::CParams` structure.
+The parameters are represented by a `CDecisionTree::CParams` structure.
 
 - *MinContinuousSubsetSize* — [for continuous features] the minimum number of vectors corresponding to a node subtree.
 - *MinDiscreteSubsetSize* — [for discrete features] the minimum number of vectors corresponding to a node subtree.
@@ -63,14 +63,14 @@ See a simple example of training a decision tree classifier. The input data set 
 ```c++
 CPtr<IDecisionTreeModel> buildModel( IProblem* data )
 {
-	CDecisionTreeTrainingModel::CParams param;
+	CDecisionTree::CParams param;
 	param.MinContinuousSubsetPart = 0.10; // Each subtree should contain no less than 10% of all nodes
 	param.MinContinuousSubsetSize = 128; // Suppose only continuous features are used
 	param.MinSplitSize = 16; // Each subtree should contain no less than 16 nodes
 	param.MaxTreeDepth = 10; // The tree depth shouldn't be more than 10
-	param.SplitCriterion = CDecisionTreeTrainingModel::SC_InformationGain;
+	param.SplitCriterion = CDecisionTree::SC_InformationGain;
 
-	CDecisionTreeTrainingModel builder( param );
+	CDecisionTree builder( param );
 
 	return dynamic_cast<IDecisionTreeModel*>( builder.Train( data ).Ptr() );
 }
