+++ NeoML/docs/ru/API/ClassificationAndRegression/README.md
@@ -43,7 +43,7 @@
 
 Метод дерева решений представляет собой классификацию на основе последовательного сравнения признаков данного объекта с некоторыми пороговыми значениями. На основе результата сравнения можно попасть в одну из дочерних вершин, и так далее, до тех пор, пока не будет достигнута листовая вершина, в которой производится окончательная классификация.
 
-Дерево решений в **NeoML** реализовано классом [CDecisionTreeTrainingModel](DecisionTree.md), а обученная им модель предоставляет интерфейс `IDecisionTreeModel`.
+Дерево решений в **NeoML** реализовано классом [CDecisionTree](DecisionTree.md), а обученная им модель предоставляет интерфейс `IDecisionTreeModel` или [`IOneVersusAllModel`](OneVersusAll.md#model) в зависимости от количества классов в датасете.
 
 ### Классификация методом один против всех
