+++ NeoML/src/CMakeLists.txt
@@ -102,7 +102,7 @@ target_sources( ${PROJECT_NAME} PRIVATE
     TraditionalML/DecisionTreeNodeClassificationStatistic.cpp
     TraditionalML/DecisionTreeNodeClassificationStatistic.h
     TraditionalML/DecisionTreeNodeStatisticBase.h
-    TraditionalML/DecisionTreeTrainingModel.cpp
+    TraditionalML/DecisionTree.cpp
     TraditionalML/DifferentialEvolution.cpp
     TraditionalML/FeatureSelection.cpp
     TraditionalML/FirstComeClustering.cpp
@@ -241,6 +241,7 @@ target_sources( ${PROJECT_NAME} PRIVATE
     ../include/NeoML/TraditionalML/CrossValidationSubProblem.h
     ../include/NeoML/TraditionalML/CommonCluster.h
     ../include/NeoML/TraditionalML/CrossValidation.h
+    ../include/NeoML/TraditionalML/DecisionTree.h
     ../include/NeoML/TraditionalML/DecisionTreeTrainingModel.h
     ../include/NeoML/TraditionalML/DifferentialEvolution.h
     ../include/NeoML/TraditionalML/FeatureSelection.h
