+++ NeoML/src/TraditionalML/DecisionTree.cpp
@@ -16,7 +16,7 @@ limitations under the License.
 #include <common.h>
 #pragma hdrstop
 
-#include <NeoML/TraditionalML/DecisionTreeTrainingModel.h>
+#include <NeoML/TraditionalML/DecisionTree.h>
 #include <DecisionTreeNodeBase.h>
 #include <DecisionTreeClassificationModel.h>
 #include <DecisionTreeNodeClassificationStatistic.h>
@@ -30,9 +30,9 @@ IDecisionTreeModel::~IDecisionTreeModel()
 
 //---------------------------------------------------------------------------------------------------------
 
-const int CDecisionTreeTrainingModel::MaxClassifyNodesCacheSize;
+const int CDecisionTree::MaxClassifyNodesCacheSize;
 
-CDecisionTreeTrainingModel::CDecisionTreeTrainingModel( const CParams& _params, CRandom* _random ) :
+CDecisionTree::CDecisionTree( const CParams& _params, CRandom* _random ) :
 	params( _params ),
 	random( _random ),
 	logStream( 0 ),
@@ -51,11 +51,11 @@ CDecisionTreeTrainingModel::CDecisionTreeTrainingModel( const CParams& _params, 
 	NeoAssert( 0.00 <= params.ConstNodeThreshold && params.ConstNodeThreshold <= 1.0 );
 }
 
-CDecisionTreeTrainingModel::~CDecisionTreeTrainingModel()
+CDecisionTree::~CDecisionTree()
 {
 }
 
-CPtr<IModel> CDecisionTreeTrainingModel::Train( const IProblem& problem )
+CPtr<IModel> CDecisionTree::Train( const IProblem& problem )
 {
 	NeoAssert( problem.GetVectorCount() > 0 );
 	NeoAssert( problem.GetClassCount() > 0 );
@@ -68,7 +68,7 @@ CPtr<IModel> CDecisionTreeTrainingModel::Train( const IProblem& problem )
 	return root.Ptr();
 }
 
-CPtr<CDecisionTreeNodeBase> CDecisionTreeTrainingModel::buildTree( int vectorCount )
+CPtr<CDecisionTreeNodeBase> CDecisionTree::buildTree( int vectorCount )
 {
 	if( logStream != 0 ) {
 		*logStream << "\nDecision tree training started:\n";
@@ -119,7 +119,7 @@ CPtr<CDecisionTreeNodeBase> CDecisionTreeTrainingModel::buildTree( int vectorCou
 }
 
 // Builds one tree level
-bool CDecisionTreeTrainingModel::buildTreeLevel( const CSparseFloatMatrixDesc& matrix, int level, CDecisionTreeNodeBase& root ) const
+bool CDecisionTree::buildTreeLevel( const CSparseFloatMatrixDesc& matrix, int level, CDecisionTreeNodeBase& root ) const
 {
 	if( logStream != 0 ) {
 		*logStream << "\nBuild level " << level << ":\n";
@@ -165,7 +165,7 @@ bool CDecisionTreeTrainingModel::buildTreeLevel( const CSparseFloatMatrixDesc& m
 
 // Gathers statistics for the nodes of one level
 // Returns true if all nodes were traversed and false if another pass is needed
-bool CDecisionTreeTrainingModel::collectStatistics( const CSparseFloatMatrixDesc& matrix, int level, CDecisionTreeNodeBase* root ) const
+bool CDecisionTree::collectStatistics( const CSparseFloatMatrixDesc& matrix, int level, CDecisionTreeNodeBase* root ) const
 {
 	NeoAssert( level > 0 );
 	NeoAssert( root != 0 );
@@ -223,7 +223,7 @@ bool CDecisionTreeTrainingModel::collectStatistics( const CSparseFloatMatrixDesc
 
 // Splits the specified node according to the accumulated statistics
 // Returns true if new nodes were created when splitting
-bool CDecisionTreeTrainingModel::split( const CDecisionTreeNodeStatisticBase& nodeStatistics, int level ) const
+bool CDecisionTree::split( const CDecisionTreeNodeStatisticBase& nodeStatistics, int level ) const
 {
 	CDecisionTreeNodeBase& node = nodeStatistics.GetNode();
 	CArray<double> predictions;
@@ -301,7 +301,7 @@ bool CDecisionTreeTrainingModel::split( const CDecisionTreeNodeStatisticBase& no
 }
 
 // Generates an array of features used
-void CDecisionTreeTrainingModel::generateUsedFeatures( int randomSelectedFeaturesCount, int featuresCount,
+void CDecisionTree::generateUsedFeatures( int randomSelectedFeaturesCount, int featuresCount,
 	CArray<int>& features ) const
 {
 	features.Empty();
@@ -324,13 +324,13 @@ void CDecisionTreeTrainingModel::generateUsedFeatures( int randomSelectedFeature
 }
 
 // Creates a node
-CPtr<CDecisionTreeNodeBase> CDecisionTreeTrainingModel::createNode() const
+CPtr<CDecisionTreeNodeBase> CDecisionTree::createNode() const
 {
 	return FINE_DEBUG_NEW CDecisionTreeClassificationModel();
 }
 
 // Creates a node statistics object
-CDecisionTreeNodeStatisticBase* CDecisionTreeTrainingModel::createStatistic( CDecisionTreeNodeBase* node ) const
+CDecisionTreeNodeStatisticBase* CDecisionTree::createStatistic( CDecisionTreeNodeBase* node ) const
 {
 	CArray<int> features;
 	generateUsedFeatures( params.RandomSelectedFeaturesCount, classificationProblem->GetFeatureCount(), features );
